# perf(internal/httputil): optimize ContentSubtype function